requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests and API client
requests==2.31.0

# Data validation and settings
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Optional: For better development experience
python-multipart==0.0.6
